move deploy_keystore_account to fixtures
