+++ crates/sncast/tests/helpers/devnet.rs
@@ -1,14 +1,11 @@
-use super::fixtures::{convert_to_hex, get_address_from_keystore, mint_token};
-use crate::helpers::constants::{DEVNET_ENV_FILE, DEVNET_OZ_CLASS_HASH, SEED, URL};
-use crate::helpers::fixtures::{declare_contract, declare_deploy_contract, remove_devnet_env};
+use crate::helpers::constants::{DEVNET_ENV_FILE, SEED, URL};
+use crate::helpers::fixtures::{declare_contract, declare_deploy_contract, remove_devnet_env, deploy_keystore_account};
 use ctor::{ctor, dtor};
-use snapbox::cmd::cargo_bin;
-use sncast::helpers::constants::KEYSTORE_PASSWORD_ENV_VAR;
 use std::net::TcpStream;
 use std::process::{Command, Stdio};
 use std::string::ToString;
 use std::time::{Duration, Instant};
-use std::{env, fs};
+use std::fs;
 use tokio::runtime::Runtime;
 use url::Url;
 
@@ -74,42 +71,6 @@ fn start_devnet() {
     rt.block_on(deploy_keystore_account());
 
     dotenv::from_filename(DEVNET_ENV_FILE).unwrap();
-}
-
-async fn deploy_keystore_account() {
-    let keystore_path = "tests/data/keystore/deployed_key.json";
-    let account_path = "tests/data/keystore/deployed_account_copy.json";
-
-    fs::copy("tests/data/keystore/deployed_account.json", account_path).unwrap();
-    env::set_var(KEYSTORE_PASSWORD_ENV_VAR, "123");
-
-    let address = get_address_from_keystore(keystore_path, account_path, KEYSTORE_PASSWORD_ENV_VAR);
-
-    mint_token(
-        &convert_to_hex(&address.to_string()),
-        9_999_999_999_999_999_999,
-    )
-    .await;
-
-    let args = vec![
-        "--url",
-        URL,
-        "--keystore",
-        keystore_path,
-        "--account",
-        account_path,
-        "account",
-        "deploy",
-        "--max-fee",
-        "99999999999999999",
-        "--class-hash",
-        DEVNET_OZ_CLASS_HASH,
-    ];
-
-    Command::new(cargo_bin!("sncast"))
-        .args(args)
-        .output()
-        .unwrap();
 }
 
 #[cfg(test)]
